refactor: centralize history component
